pre release webservice full prod dict buys slot feeds on linux

+++ banta/packages/optional/buys.py
@@ -25,7 +25,7 @@ class Buys(_pack.GenericModule):
 		self.app.window.acNewStockBuy.setEnabled(True)
 		self.app.window.acNewStockBuy.triggered.connect(self.newBuy)
 		self.dialog.cbProd.currentIndexChanged[int].connect(self.prodChanged)
-		self.dialog.dPacks.valueChanged[float].connect(self.packChanged)
+		self.dialog.dPacks.valueChanged[int].connect(self.packChanged)
 		prod_mod_name = banta.packages.base.products.Products.NAME
 		model = self.app.modules[prod_mod_name].model
 		cbp = self.dialog.cbProd
